pyproject.toml
[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[project]
name = "maxwell-daemon"
version = "0.1.0"
description = "Multi-backend autonomous code agent orchestrator with fleet management"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "D-sorganization", email = "dieterolson@gmail.com" }]
keywords = ["ai", "agents", "llm", "automation", "orchestration", "fleet"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "httpx>=0.27.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "pyyaml>=6.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=13.0",
    "prometheus-client>=0.20.0",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
ollama = ["ollama>=0.3.0"]
google = ["google-cloud-aiplatform>=1.70.0"]
azure = ["azure-identity>=1.18.0"]
grpc = ["grpcio>=1.66.0", "grpcio-tools>=1.66.0", "protobuf>=5.28.0"]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "respx>=0.21.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "opentelemetry-sdk>=1.30.0",
    "types-pyyaml",
]
all = ["maxwell-daemon[ollama,google,azure,grpc,dev]"]

[project.scripts]
maxwell-daemon = "maxwell_daemon.cli.main:app"
maxwell-daemon-runner = "maxwell_daemon.daemon.runner:main"

[project.urls]
Homepage = "https://github.com/D-sorganization/Maxwell-Daemon"
Documentation = "https://github.com/D-sorganization/Maxwell-Daemon#readme"
Repository = "https://github.com/D-sorganization/Maxwell-Daemon"
Issues = "https://github.com/D-sorganization/Maxwell-Daemon/issues"

[tool.hatch.build.targets.wheel]
packages = ["maxwell_daemon"]

[tool.hatch.build.targets.wheel.force-include]
"maxwell_daemon/api/ui/index.html" = "maxwell_daemon/api/ui/index.html"
"maxwell_daemon/api/ui/app.js" = "maxwell_daemon/api/ui/app.js"
"maxwell_daemon/api/ui/style.css" = "maxwell_daemon/api/ui/style.css"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers --cov=maxwell_daemon --cov-report=term-missing"

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["prometheus_client", "prometheus_client.*", "opentelemetry.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["anthropic", "anthropic.*", "openai", "openai.*"]
ignore_missing_imports = false

[tool.pydantic-mypy]
init_forbid_extra = false
init_typed = true
warn_required_dynamic_aliases = true

[tool.bandit]
# Tests contain intentional assertions and fake secrets for fixtures — skip.
exclude_dirs = ["tests", "docs"]
# B101 is assert-removed-under-optimisation, which is a style nit not a vuln.
# B404/B603 are about subprocess — we deliberately use subprocess with argv lists.
skips = ["B101", "B404", "B603"]
